% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoSleepR.R
\name{nosleep_on}
\alias{nosleep_on}
\title{Turn nosleep on}
\usage{
nosleep_on(keep_display = TRUE)
}
\arguments{
\item{keep_display}{logical. If TRUE, also prevent the display from sleeping.
the nosleep state is automatically released.}
}
\description{
Prevent the system from going to sleep while R code is running.
}
