% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoSleepR-package.R
\docType{package}
\name{NoSleepR}
\alias{NoSleepR-package}
\alias{NoSleepR}
\title{NoSleepR: Prevent system sleep during long R tasks}
\description{
NoSleepR provides a cross-platform interface to prevent the operating system
from going to sleep while long-running R code is executing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hetalang/NoSleepR}
  \item Report bugs at \url{https://github.com/hetalang/NoSleepR/issues}
}

}
\author{
\strong{Maintainer}: Evgeny Metelkin \email{evgeny.metelkin@gmail.com}

}
\keyword{internal}
