% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoSleepR.R
\name{with_nosleep}
\alias{with_nosleep}
\title{Execute an expression while preventing the system from sleeping}
\usage{
with_nosleep(expr, keep_display = TRUE)
}
\arguments{
\item{expr}{Expression to execute while nosleep is on.}

\item{keep_display}{logical. If TRUE, also prevent the display from sleeping.}
}
\value{
The result of evaluating \code{expr}.
}
\description{
Execute an expression while preventing the system from sleeping
}
